backend/requirements.txt
fastapi
uvicorn
pymongo[srv]
python-dotenv
fastapi-cors
passlib[bcrypt]
python-jose[cryptography]
pydantic